requirements.txt
dnspython==1.16.0
ecdsa==0.15
pycryptodome==3.9.7
six==1.14.0
